FREEI-813 Properly scope tokens
When no scope is provided for an application we need to make sure
that we grant the oauth jwt token access to both the GraphQL and REST
APIs. Likewise, if no scope is defined on the request to fetch a token
we need to grab the scope that is set for the application use that
value instead. Both these changes make dealing with the GraphQL and
REST apis a little more consistent.