Implements an API Base Test Class
This PR creates an API Base Test Class to assist with creating unit
tests for GraphQL APIs by utilizing the slim testing framework. The
helper also generates a unit test application inside of FreePBX and
creates a token that can be used for authenticaiton.

To use this helper, simple include the ApiBaseTestCase.php file into
a test file and extend the ApiBaseTestCase class.